# backend/auction/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auctions": {
            "get": {
                "description": "获取竞拍列表，支持按状态筛选和分页",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "获取竞拍列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "竞拍状态：0-待开始, 1-进行中, 2-已结束, 3-已取消",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的竞拍场次",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "创建竞拍",
                "parameters": [
                    {
                        "description": "竞拍信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateAuctionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Auction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auctions/{id}": {
            "get": {
                "description": "获取指定竞拍的详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "获取竞拍详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "竞拍ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Auction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auctions/{id}/bids": {
            "get": {
                "description": "获取指定竞拍的所有出价记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "获取竞拍出价记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "竞拍ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "返回数量限制",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Bid"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "在指定竞拍中出价",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bid"
                ],
                "summary": "出价竞拍",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "竞拍ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "出价信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PlaceBidRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.PlaceBidResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auctions/{id}/cancel": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "取消指定的竞拍场次",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "取消竞拍",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "竞拍ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auctions/{id}/ranking": {
            "get": {
                "description": "获取指定竞拍的出价排名列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bid"
                ],
                "summary": "获取竞拍排名",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "竞拍ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "返回数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auctions/{id}/result": {
            "get": {
                "description": "获取指定竞拍的最终结果",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auction"
                ],
                "summary": "获取竞拍结果",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "竞拍ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取用户的通知列表，支持分页和筛选",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "获取通知列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "仅未读通知",
                        "name": "unread_only",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.NotificationListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/notifications/read-all": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "将用户的所有未读通知标记为已读",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "标记所有通知为已读",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/notifications/unread-count": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取用户的未读通知总数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "获取未读通知数量",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UnreadCountResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/notifications/{id}/read": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "将指定通知标记为已读",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "标记通知为已读",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "通知ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.CreateAuctionRequest": {
            "type": "object",
            "required": [
                "duration",
                "product_id"
            ],
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "increment": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                },
                "start_price": {
                    "type": "number"
                }
            }
        },
        "handler.PlaceBidRequest": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "user_id": {
                    "description": "用于测试，生产环境应删除",
                    "type": "integer"
                }
            }
        },
        "model.Auction": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "creator_id": {
                    "description": "竞拍创建者ID（主播）",
                    "type": "integer"
                },
                "current_price": {
                    "type": "number"
                },
                "delay_used": {
                    "description": "已延时秒数",
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "live_stream_id": {
                    "description": "直播间ID（新增字段）",
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.AuctionStatus"
                },
                "winner_id": {
                    "type": "integer"
                }
            }
        },
        "model.AuctionStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-comments": {
                "AuctionStatusCancelled": "已取消",
                "AuctionStatusDelayed": "延时中",
                "AuctionStatusEnded": "已结束",
                "AuctionStatusOngoing": "进行中",
                "AuctionStatusPending": "待开始"
            },
            "x-enum-descriptions": [
                "待开始",
                "进行中",
                "延时中",
                "已结束",
                "已取消"
            ],
            "x-enum-varnames": [
                "AuctionStatusPending",
                "AuctionStatusOngoing",
                "AuctionStatusDelayed",
                "AuctionStatusEnded",
                "AuctionStatusCancelled"
            ]
        },
        "model.Bid": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "auction_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.JSONMap": {
            "type": "object",
            "additionalProperties": true
        },
        "model.Notification": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/model.JSONMap"
                },
                "id": {
                    "type": "integer"
                },
                "read_at": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/model.NotificationType"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.NotificationListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Notification"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.NotificationType": {
            "type": "string",
            "enum": [
                "bid_outbid",
                "auction_won",
                "auction_lost",
                "order_paid",
                "order_shipped",
                "order_completed",
                "new_product",
                "auction_starting",
                "product_unpublished",
                "auction_ended"
            ],
            "x-enum-comments": {
                "NotificationTypeAuctionEnded": "竞拍已结束",
                "NotificationTypeAuctionLost": "竞拍未中标",
                "NotificationTypeAuctionStarting": "竞拍即将开始",
                "NotificationTypeAuctionWon": "竞拍中标",
                "NotificationTypeBidOutbid": "出价被超越",
                "NotificationTypeNewProduct": "新商品发布",
                "NotificationTypeOrderCompleted": "订单已完成",
                "NotificationTypeOrderPaid": "订单已支付",
                "NotificationTypeOrderShipped": "订单已发货",
                "NotificationTypeProductUnpublished": "商品已下架"
            },
            "x-enum-descriptions": [
                "出价被超越",
                "竞拍中标",
                "竞拍未中标",
                "订单已支付",
                "订单已发货",
                "订单已完成",
                "新商品发布",
                "竞拍即将开始",
                "商品已下架",
                "竞拍已结束"
            ],
            "x-enum-varnames": [
                "NotificationTypeBidOutbid",
                "NotificationTypeAuctionWon",
                "NotificationTypeAuctionLost",
                "NotificationTypeOrderPaid",
                "NotificationTypeOrderShipped",
                "NotificationTypeOrderCompleted",
                "NotificationTypeNewProduct",
                "NotificationTypeAuctionStarting",
                "NotificationTypeProductUnpublished",
                "NotificationTypeAuctionEnded"
            ]
        },
        "model.UnreadCountResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                }
            }
        },
        "service.PlaceBidResult": {
            "type": "object",
            "properties": {
                "current_price": {
                    "type": "number"
                },
                "message": {
                    "type": "string"
                },
                "rank": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                },
                "winner_id": {
                    "type": "integer"
                }
            }
        }
    }
}